# static/specs/openapi/catalog/catalog_agent.yaml
openapi: 3.0.3
info:
  title: Rainbow Catalog Agent API - Entity Management
  description: OpenAPI specification for the Rainbow Catalog Agent Entity Management endpoints.
  version: 1.0.0
servers:
  - url: /v1/catalog-agent
    description: Base path for the catalog agent API
...
security:
  - BearerAuth: []
paths:
...
  /datasets/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags:
        - Datasets
      summary: Get dataset by ID
      operationId: getDatasetById
      responses:
        "200":
          description: Dataset found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - Datasets
      summary: Update dataset by ID
      operationId: updateDatasetById
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditDatasetDto"
      responses:
        "200":
          description: Dataset updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Datasets
      summary: Delete dataset by ID
      operationId: deleteDatasetById
      responses:
        "202":
          description: Dataset deleted (Accepted)
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # --- Distributions ---
...
  /policy-templates:
    get:
      tags:
        - Policy Templates
      summary: Get all policy templates
      operationId: getAllPolicyTemplates
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/PageParam"
      responses:
        "200":
          description: List of policy templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PolicyTemplateDto"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Policy Templates
      summary: Create policy template
      operationId: createPolicyTemplate
      parameters:
        - name: silent
          in: query
          schema:
            type: boolean
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewPolicyTemplateDto"
            example:
              id: "my-template"
              version: "1.0"
              content:
                permission:
                  - action: "use"
      responses:
        "200":
          description: Policy template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyTemplateDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "406":
          description: Not acceptable (Silent mode error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /policy-templates/batch:
    post:
      tags:
        - Policy Templates
      summary: Get batch policy templates
      operationId: getBatchPolicyTemplates
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchRequestsAsString"
      responses:
        "200":
          description: List of policy templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PolicyTemplateDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /policy-templates/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags:
        - Policy Templates
      summary: Get policy template by ID (all versions)
      operationId: getPolicyTemplateById
      responses:
        "200":
          description: List of policy templates (versions)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PolicyTemplateDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /policy-templates/{id}/{version}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
      - name: version
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Policy Templates
      summary: Get policy template by ID and version
      operationId: getPolicyTemplateByIdAndVersion
      responses:
        "200":
          description: Policy template found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyTemplateDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Policy Templates
      summary: Delete policy template by ID and version
      operationId: deletePolicyTemplateByIdAndVersion
      responses:
        "202":
          description: Policy template deleted (Accepted)
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /policy-templates/instantiate-odrl-offer:
    post:
      tags:
        - Policy Templates
      summary: Instantiate ODRL offer from template
      operationId: instantiateOdrlOffer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewPolicyInstantiationDto"
      responses:
        "202":
          description: Instantiated (Accepted)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OdrlPolicyDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # --- Peer Catalogs ---
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        format: int64
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        format: int64
...
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
  schemas:
...
    ErrorInfo:
      type: object
      required:
        - message
        - error_code
        - cause
      properties:
        message:
          type: string
        error_code:
          type: integer
        details:
          type: string
        cause:
          type: string
      example:
        message: "Resource not found"
        error_code: 3200
        status_code: 404
        cause: "Catalog with ID urn:uuid:123 not found"

    # --- Common Wrappers ---
...
    BatchRequestsAsString:
      type: object
      properties:
        ids:
          type: array
          items:
            type: string

    # --- Catalogs ---
...
    DatasetDto:
      type: object
      properties:
        id:
          type: string
          format: urn
        dctConformsTo:
          type: string
        dctCreator:
          type: string
        dctIdentifier:
          type: string
        dctIssued:
          type: string
          format: date-time
        dctModified:
          type: string
          format: date-time
        dctTitle:
          type: string
        dctDescription:
          type: string
        catalogId:
          type: string
          format: urn
      example:
        {
          "catalogId": "urn:catalog:4958f37d-17cf-44e7-ae18-1f26b6fe15ff",
          "dctConformsTo": null,
          "dctCreator": null,
          "dctDescription": null,
          "dctIdentifier": "urn:dataset:10825ca3-a75e-43bd-9a82-603759cb8dec",
          "dctIssued": "2026-01-10T14:23:29.091150Z",
          "dctModified": null,
          "dctTitle": "superfoaf",
          "id": "urn:dataset:10825ca3-a75e-43bd-9a82-603759cb8dec"
        }
...
    EditDatasetDto:
      type: object
      properties:
        dctConformsTo:
          type: string
        dctCreator:
          type: string
        dctTitle:
          type: string
        dctDescription:
          type: string
      example:
        {
          "dctConformsTo": null,
          "dctCreator": null,
          "dctDescription": null,
          "dctTitle": "superfoaf",
        }

    # --- Distributions ---
...
    OdrlPolicyDto:
      type: object
      properties:
        id:
          type: string
          format: urn
        odrlOffer:
          $ref: "#/components/schemas/OdrlPolicyInfo"
        entity:
          type: string
          format: urn
        entityType:
          type: string
          enum: [Distribution, DataService, Catalog, Dataset]
        createdAt:
          type: string
          format: date-time
        sourceTemplateId:
          type: string
        sourceTemplateVersion:
          type: string
        instantiationParameters:
          type: object
      example:
        {
          "createdAt": "2026-01-10T14:40:30.978559Z",
          "entity": "urn:catalog:35d10981-1d0b-4204-b32e-847b827ce635",
          "entityType": "Catalog",
          "id": "urn:odrl-policy:abad083e-78a8-4f90-aaca-7c5d6af6c251",
          "instantiationParameters": {
            "$bla": "hola"
          },
          "odrlOffer": {
            "permission": [
              {
                "action": "use",
                "constraint": [
                  {
                    "leftOperand": "bla",
                    "operator": "eq",
                    "rightOperand": "hola"
                  }
                ]
              }
            ]
          },
          "sourceTemplateId": "policy-1",
          "sourceTemplateVersion": "1.0"
        }
...
    OdrlPolicyInfo:
      type: object
      properties:
        profile:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        permission:
          type: array
          items:
            $ref: "#/components/schemas/OdrlPermission"
        obligation:
          type: array
          items:
            $ref: "#/components/schemas/OdrlObligation"
        prohibition:
          type: array
          items:
            $ref: "#/components/schemas/OdrlObligation"
    OdrlPermission:
      type: object
      required:
        - action
      properties:
        action:
          type: string
        constraint:
          type: array
          items:
            $ref: "#/components/schemas/OdrlConstraint"
        duty:
          $ref: "#/components/schemas/OdrlDuty"
    OdrlDuty:
      type: object
      required:
        - action
      properties:
        action:
          type: string
        constraint:
          type: array
          items:
            $ref: "#/components/schemas/OdrlConstraint"
    OdrlObligation:
      $ref: "#/components/schemas/OdrlDuty"
    OdrlConstraint:
      oneOf:
        - $ref: "#/components/schemas/OdrlAtomicConstraint"
        - $ref: "#/components/schemas/OdrlLogicalConstraint"
    OdrlAtomicConstraint:
      type: object
      required:
        - rightOperand
        - leftOperand
        - operator
      properties:
        leftOperand:
          type: string
        operator:
          type: string
          enum:
            [
              eq,
              gt,
              gteq,
              lteq,
              hasPart,
              isA,
              isAllOf,
              isAnyOf,
              isNoneOf,
              isPartOf,
              lt,
              termLteq,
              neq,
            ]
        rightOperand:
          oneOf:
            - type: string
            - type: object
            - type: array
              items: {}
    OdrlLogicalConstraint:
      type: object
      properties:
        and:
          type: array
          items:
            $ref: "#/components/schemas/OdrlConstraint"
        andSequence:
          type: array
          items:
            $ref: "#/components/schemas/OdrlConstraint"
        or:
          type: array
          items:
            $ref: "#/components/schemas/OdrlConstraint"
        xone:
          type: array
          items:
            $ref: "#/components/schemas/OdrlConstraint"

    # --- Policy Templates ---
    PolicyTemplateDto:
      type: object
      properties:
        id:
          type: string
        version:
          type: string
        date:
          type: string
          format: date-time
        title:
          $ref: "#/components/schemas/LocalizedText"
        description:
          $ref: "#/components/schemas/LocalizedText"
        author:
          type: string
        content:
          $ref: "#/components/schemas/OdrlPolicyInfo"
        parameters:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ParameterDefinition"
      example:
        {
          "author": "",
          "content": {
            "permission": [
              {
                "action": "use",
                "constraint": [
                  {
                    "leftOperand": "bla",
                    "operator": "eq",
                    "rightOperand": "$bla"
                  }
                ]
              }
            ]
          },
          "date": "2026-01-09T13:25:40.004265Z",
          "description": "Mi example super policy",
          "id": "policy-1",
          "parameters": {
            "$bla": {
              "dataType": "string",
              "restrictions": {},
              "ui": {
                "hidden": false
              }
            }
          },
          "title": "Mi example policy",
          "version": "1.0"
        }
    NewPolicyTemplateDto:
      type: object
      required:
        - content
      properties:
        id:
          type: string
        version:
          type: string
        date:
          type: string
          format: date-time
        title:
          $ref: "#/components/schemas/LocalizedText"
        description:
          $ref: "#/components/schemas/LocalizedText"
        author:
          type: string
        content:
          $ref: "#/components/schemas/OdrlPolicyInfo"
        parameters:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ParameterDefinition"
      example:
        {
          "author": "",
          "content": {
            "permission": [
              {
                "action": "use",
                "constraint": [
                  {
                    "leftOperand": "bla",
                    "operator": "eq",
                    "rightOperand": "$bla"
                  }
                ]
              }
            ]
          },
          "date": "2026-01-09T13:25:40.004265Z",
          "description": "Mi example super policy",
          "id": "policy-1",
          "parameters": {
            "$bla": {
              "dataType": "string",
              "restrictions": {},
              "ui": {
                "hidden": false
              }
            }
          },
          "title": "Mi example policy",
          "version": "1.0"
        }
    NewPolicyInstantiationDto:
      type: object
      required:
        - id
        - version
        - parameters
        - entityId
        - entityType
      properties:
        id:
          type: string
        version:
          type: string
        parameters:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/PolicyTemplateAllowedDefaultValues"
        entityId:
          type: string
          format: urn
        entityType:
          type: string
          enum: [Distribution, DataService, Catalog, Dataset]
      example:
        {
          "id": "policy-1",
          "version": "1.0",
          "parameters": {
            "$bla": "hola"
          },
          "entityId": "urn:catalog:35d10981-1d0b-4204-b32e-847b827ce635",
          "entityType": "Catalog"
        }
    LocalizedText:
      type: object
      additionalProperties:
        type: string
    ParameterDefinition:
      type: object
      properties: {}
    PolicyTemplateAllowedDefaultValues:
      oneOf:
        - type: string
        - type: number

    # --- DSP Types (Needed for PeerCatalog) ---
...